pyproject.toml
[project]
name = "agentic-cli"
version = "0.1.0"
description = "Agentic CLI for computer control via semantic commands"
requires-python = ">=3.11"
dependencies = [
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "setuptools>=82.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
